Python sample programs1
Some python programs including program to read the sides of a triangle a, b and calculate the hypotenue using the formula square root(square of a + square of b), to read a string and use appropriate string operation to print the string 20 times in a line with no space, read a string and find the number of vowels and consonants in the string, print the multiplication table of n, 2<=n<=10. Replace the 'while' statement with an equivalent 'for'  statement to get the same output, create a list of strings. Separate palindromes in the list to another list using reverse() function